Refactor nginx configuration and add ACME location and cert paths
This commit refactors the ODK sync locations into a separate config
file which is included in both the http and https server
configurations. This should help prevent the dev http and prod https
configurations from drifting out of sync over time.

This commit also adds a location to the http server on port 80, in the
https configuration, to allow ACME challenges only to still go out
over http. These challenges are secured over-the-top and need to work
to allow https certificates to be acquired.

Lastly, again in the https configuration, this commit lays the ground
work for moving the certificate locations to a standardized path in
https deployment configurations.