Refactor https configuration to own layer and add certbot
This commit adds automatic certificate provisioning and renewal for
https. It uses certbot to automatically get new certificates from
letsencrypt. The https configration is somewhat complicated, so this
commit refactors it to a separate docker-compose layer which is
applied on top of the existing compose file.

As an example, with this commit the default https stack can be brought
up with `sudo docker stack deploy -c docker-compose.yml -c
docker-compose-https.yml syncldap`