Fix certbot renewal in field systems
The test setup used for the initial integration of certbot renewal had
additional metadata present that allowed the renew command to complete
successfully. In production after a minimal bootstrap this metadata is
not available. This commit updates the renewal flow to generate a new
"staging" certificate, which is then copied over to the production
directory if it's updated successfully. This allows certbot to
generate the metadata required to update the certificate in the
future. The certonly command works for both the initial allocation of
a new certificate as well as the update of an existing certificate.

This commit was tested on a clean cloud host (azure), and validated
that an update occurs when needed while not executing extra queries
to the CA when the staging certificate is present and not due for
renewal.